fix(ImagePreview): add test cases
